+++ bot/main.py
@@ -1,7 +1,7 @@
 import telebot
 import os
 from telebot.storage import StateMemoryStorage
-from funcs import add_moder
+from funcs import add_moder, ban_user, unban_user
 from funcs import start, help, place, user_settings, change_distance, v1
 from ya_ai_xd import handle_location, create_navigation_keyboard
 from settings_requests import get_db_connection, get_user_status, upd_user_status
@@ -13,7 +13,7 @@ from settings_requests import get_user_request_ids, get_current_index, upd_curre
     get_user_request_comment_ids, upd_current_comment_index, get_current_comment_index
 from comments import create_comment_card, create_navigation_keyboard_for_comments, get_comments, \
     get_user_comments, create_navigation_keyboard_for_user_comments, print_place
-from funcs import set_comment, set_rating
+from funcs import set_comment, set_rating, check_banned
 from commet_requests import delete_comment
 from users_requests import get_user_role
 from time import sleep
@@ -49,24 +49,41 @@ def stop(message):
 @tb.message_handler(commands=["start"])
 def start_handler(message):
     """Чистый запуск бота"""
-    start(message)
+    if (check_banned(message.from_user.id) == False):
+        start(message)
+    else:
+        tb.sent_message(message.from_user.id, "Вас занесли в черный список, чтобы обжаловать решение пишите:\n"
+                        "@hom1e101\n@flovvey36")
 
 
 @tb.message_handler(commands=["help"])
 def help_handler(message):
     """Помощь пользователю"""
-    help(message)
+    if (check_banned(message.from_user.id) == False):
+        help(message)
+    else:
+        tb.sent_message(message.from_user.id, "Вас занесли в черный список, чтобы обжаловать решение пишите:\n"
+                        "@hom1e101\n@flovvey36")
 
 
 @tb.message_handler(commands=["settings"])
 def settings_handler(message):
     """Настройки пользователя"""
-    user_settings(message)
+    if (check_banned(message.from_user.id) == False):
+        user_settings(message)
+    else:
+        tb.sent_message(message.from_user.id, "Вас занесли в черный список, чтобы обжаловать решение пишите:\n"
+                        "@hom1e101\n@flovvey36")
+    
 
 
 @tb.message_handler(commands=["promote"])
 def promotion(message):
     """Повышение роли пользователя"""
+    if (check_banned(message.from_user.id) == True):
+        tb.sent_message(message.from_user.id, "Вас занесли в черный список, чтобы обжаловать решение пишите:\n"
+                        "@hom1e101\n@flovvey36")
+        return
     tb.delete_message(message.from_user.id, message.id)
     sent_message = tb.send_message(message.from_user.id, "Получил вашу заявку, в ближайшее время администраторы ее рассмотрят")
     sleep(1.5)
@@ -85,6 +102,9 @@ def promotion(message):
 @tb.message_handler(commands=["add_moder"])
 def add_moder_handler(message):
     """Добавление модератора администратором"""
+    if (check_banned(message.from_user.id) == True):
+        tb.sent_message(message.from_user.id, "Вас занесли в черный список, чтобы обжаловать решение пишите:\n"
+                        "@hom1e101\n@flovvey36")
     user_id = message.from_user.id
     tb.delete_message(user_id, message.id)
     with get_db_connection() as conn:
@@ -98,9 +118,50 @@ def add_moder_handler(message):
             tb.delete_message(user_id, sent_message.id)
 
 
+@tb.message_handler(commands=["ban_user"])
+def ban_user_handler(message):
+    """удаление пользователя администратором"""
+    if (check_banned(message.from_user.id) == True):
+        tb.sent_message(message.from_user.id, "Вас занесли в черный список, чтобы обжаловать решение пишите:\n"
+                        "@hom1e101\n@flovvey36")
+    user_id = message.from_user.id
+    tb.delete_message(user_id, message.id)
+    with get_db_connection() as conn:
+        if get_user_role(conn, user_id) in ["admin", "moderator"]:
+            upd_user_status(conn, user_id, "ban_user")
+            sent_message = tb.send_message(message.from_user.id, "Напиши username пользователя, "
+                "которого забанить в формате: @username")
+        else:
+            sent_message = tb.send_message(message.from_user.id, "Ошибка: ты не администратор и не модератор")
+            sleep(1)
+            tb.delete_message(user_id, sent_message.id)
+
+
+@tb.message_handler(commands=["unban_user"])
+def ban_user_handler(message):
+    """разбан пользователя администратором"""
+    if (check_banned(message.from_user.id) == True):
+        tb.sent_message(message.from_user.id, "Вас занесли в черный список, чтобы обжаловать решение пишите:\n"
+                        "@hom1e101\n@flovvey36")
+    user_id = message.from_user.id
+    tb.delete_message(user_id, message.id)
+    with get_db_connection() as conn:
+        if get_user_role(conn, user_id) in ["admin", "moderator"]:
+            upd_user_status(conn, user_id, "unban_user")
+            sent_message = tb.send_message(message.from_user.id, "Напиши username пользователя, "
+                "которого разбанить в формате: @username")
+        else:
+            sent_message = tb.send_message(message.from_user.id, "Ошибка: ты не администратор и не модератор")
+            sleep(1)
+            tb.delete_message(user_id, sent_message.id)
+
+
 @tb.message_handler()
 def message_handler(message):
     """Изменение состояния пользователя"""
+    if (check_banned(message.from_user.id) == True):
+        tb.sent_message(message.from_user.id, "Вас занесли в черный список, чтобы обжаловать решение пишите:\n"
+                        "@hom1e101\n@flovvey36")
     status = ""
     user_id = message.from_user.id
     with get_db_connection() as conn:
@@ -114,28 +175,41 @@ def message_handler(message):
         change_distance(message)
     elif status == "comments":
         change_distance(message)
-    elif status in {"r_1", "r_2", "r_3", "r_4", "r_5"}:
+    elif status in {"r_1", "r_2", "r_3", "r_4", "r_5"} or status.startswith("redact_rating_"):
         set_rating(message)
-    elif status in {"c_1", "c_2", "c_3", "c_4", "c_5"}:
+    elif status in {"c_1", "c_2", "c_3", "c_4", "c_5"} or status.startswith("redact_text_"):
         set_comment(message)
     elif status == "add_moder":
         add_moder(message)
+    elif status == "ban_user":
+        ban_user(message)
+    elif status == "unban_user":
+        unban_user(message)
 
 
 @tb.message_handler(content_types=["location"])
 def location_handler(message):
     """Обработчик для поиска мест рядом"""
+    if (check_banned(message.from_user.id) == True):
+        tb.sent_message(message.from_user.id, "Вас занесли в черный список, чтобы обжаловать решение пишите:\n"
+                        "@hom1e101\n@flovvey36")
     handle_location(message)
 
 
 @tb.message_handler(commands=['v1'])
 def gab(message):
+    if (check_banned(message.from_user.id) == True):
+        tb.sent_message(message.from_user.id, "Вас занесли в черный список, чтобы обжаловать решение пишите:\n"
+                        "@hom1e101\n@flovvey36")
     v1(message)
 
 
 @tb.callback_query_handler()
 def handle_navigation(call):
     """Обработка всех звонков инлайн кнопок"""
+    if (check_banned(call.from_user.id) == True):
+        tb.sent_message(call.from_user.id, "Вас занесли в черный список, чтобы обжаловать решение пишите:\n"
+                        "@hom1e101\n@flovvey36")
     if call.data.startswith(('prev_', 'next_', 'rate_', 'comment_', "back_", "get_comm_")):
         """Обрабатывает навигацию между местами"""
 
@@ -212,7 +286,7 @@ def handle_navigation(call):
         except Exception as e:
             tb.answer_callback_query(call.id, "Произошла ошибка. Попробуйте снова.")
 
-    elif call.data.startswith(("comm_prev_", "comm_next_", "comm_back_")):
+    elif call.data.startswith(("comm_prev_", "comm_next_", "comm_back_", "redact_")):
         """Навигация в комментариях"""
         try:
             user_id = call.from_user.id
@@ -235,6 +309,34 @@ def handle_navigation(call):
             elif call.data.startswith('comm_back_'):
                 new_index = current_index
             else:
+                if call.data.startswith('redact_text_'):
+                    comment_id = comment_ids[current_index]
+                    with get_db_connection() as conn:
+                        comment = get_comment_by_comment_id(conn, comment_id)
+                    place_id = comment["place_id"]
+                    name = ''
+                    with get_db_connection() as conn:
+                        name = get_place_by_id(conn, place_id)['name']
+
+                    tb.answer_callback_query(call.id, f"Вы хотите редактировать комментарий к месту: {name}")
+
+                    with get_db_connection() as conn:
+                        upd_user_status(conn, user_id, f'redact_text_{comment_id}')
+                    tb.send_message(user_id, "Напишите комментарий, который хотите оставить")
+                else:
+                    comment_id = comment_ids[current_index]
+                    with get_db_connection() as conn:
+                        comment = get_comment_by_comment_id(conn, comment_id)
+                    place_id = comment["place_id"]
+                    name = ''
+                    with get_db_connection() as conn:
+                        name = get_place_by_id(conn, place_id)["name"]
+
+                    tb.answer_callback_query(call.id, f"Вы хотите поменять оценку на место: {name}")
+
+                    with get_db_connection() as conn:
+                        upd_user_status(conn, user_id, f'redact_rating_{comment_id}')
+                    tb.send_message(user_id, "Напишите оценку, которую хотите оставить")
                 return
 
             with get_db_connection() as conn:
